Switch to SurfaceView for performance; Fix dropping items when it shouldn't be the case;
