+++ app_pojavlauncher/src/main/res/layout/activity_basemain.xml
@@ -20,7 +20,7 @@
 			android:layout_width="match_parent"
 			android:layout_height="match_parent">
 
-			<net.kdt.pojavlaunch.MinecraftGLView
+			<net.kdt.pojavlaunch.MinecraftGLSurfaceView
 				android:id="@+id/main_game_render_view"
 				android:layout_width="fill_parent"
 				android:layout_height="fill_parent"/>
